Removed image recommendation and renamed as matching engine
